Refactor organization sync into shared org helpers
