+++ api/src/organization.py
@@ -1,3 +1,5 @@
+import src.db as db
+from typing import Final
 from pydantic import BaseModel
 
 
@@ -13,3 +15,16 @@ class OrganizationSettings(BaseModel):
 
 
 org = OrganizationSettings()
+
+ORGANIZATION_SETTINGS_KEYS: Final[tuple[str, ...]] = tuple(
+    OrganizationSettings.model_fields.keys()
+)
+
+
+async def organization_settings_payload() -> dict[str, str]:
+    payload: dict[str, str] = {}
+    for key in ORGANIZATION_SETTINGS_KEYS:
+        setting = await db.settings.get(key, app_id=None)
+        payload[key] = setting.value if setting is not None else ''
+
+    return payload
